Backend: Initial flask App. and Database Tables.
Created the flask app in `app.py` where it has
imported the `models` module that consists of
the database classes (tables/entities) and used
the jinja2 template `general_strategy.html` to
render the hardcoded data until data are populated
in the Strategize database.

for environment preparation we uploaded used
`install_env` bash script and the to install the
MySQL server and its connector we used
`install_mysql` bash script.

To run the flask app, we uploaded `run_flask` bash
script.